+++ demo/input.html
@@ -10,7 +10,7 @@
 
         <script type="text/javascript" src="../dist/input.js"></script>
         <script type="text/javascript">
-            new InputInlineAttachment(document.getElementById("code")).bind();
+            new InputInlineAttachment(document.getElementById("code"));
         </script>
     </body>
 </html>
