+++ src/codemirror-4/codemirror-4.js
@@ -1,7 +1,7 @@
 /*jslint newcap: true */
 /*global inlineAttachment: false */
 
-import InlineAttachment from "./inline-attachment";
+import InlineAttachment from "../inline-attachment";
 
 export default class CodeMirror4InlineAttachment {
