+++ src/input/main.js
@@ -1,0 +1,14 @@
+import InputInlineAttachment from './input';
+
+(function() {
+
+  let root = this;
+
+  root.InputInlineAttachment = InputInlineAttachment;
+
+  if (typeof exports !== "undefined") {
+    exports.InputInlineAttachment = root.InputInlineAttachment;
+  }
+
+
+}).call(typeof window !== 'undefined' ? window : this);
